Abstracted master and server classes
