+++ overkill/servers/__init__.py
@@ -1,0 +1,3 @@
+"""Servers galore"""
+
+__all__ = [s for s in dir() if not s.startswith('_')]
