+++ overkill/servers/worker.py
@@ -1,91 +1,16 @@
 """This module acts as the main class to create and manage a worker server"""
 
-from base64 import encode
-from inspect import trace
 import logging
-import multiprocessing
-import socketserver
 import threading
-import traceback
-from typing import Dict, List, Tuple
+from typing import Tuple
 
-from overkill.utils.server_data_classes import _MasterInfo
-from overkill.utils.server_exceptions import AskTypeNotFoundError, ServerNotStartedError, WorkError
-from overkill.utils.server_messaging_standards import (_NEW_CONNECTION, _REJECT, _ACCEPT,
-                                                       _ACCEPT_WORK, _CLOSE_CONNECTION,
-                                                       _DELEGATE_WORK, _WORK_ERROR)
-from overkill.utils.utils import _recv_msg, _send_message, _encode_dict, _decode_message
+from overkill.servers.utils.server_exceptions import (
+    ServerAlreadyStartedError, ServerNotStartedError)
+from overkill.servers.utils.server_messaging_standards import _NEW_CONNECTION
+from overkill.servers.utils.utils import _encode_dict, _send_message
 
-_master = None  # master info
-_id = None  # worker id
-
-
-class _WorkerServer(socketserver.BaseRequestHandler):
-
-    def __init__(self, request, client_address, server) -> None:
-        socketserver.BaseRequestHandler.__init__(
-            self, request, client_address, server)
-
-    def server_close(self):
-        pass
-
-    def handle(self):
-        """Handle incoming connections
-
-        :raises Exception: internal error in master server
-        :raises askTypeNotFoundError: occurs when there is no case for an ask type
-        """
-        global _master, _id
-        data = _recv_msg(self.request)
-        try:
-            ask = _decode_message(data)
-            logging.info(ask)
-            if ask["type"] == _REJECT:
-                raise Exception("Worker rejected")
-            elif ask["type"] == _ACCEPT:
-                address = ask["master_address"]
-                _master = _MasterInfo(address)
-                logging.info(f"Master information: {_master}")
-                _id = ask["id"]
-            elif ask["type"] == _DELEGATE_WORK:
-                results = self.do_work(ask)
-                _send_message(_encode_dict(
-                    {"type": _ACCEPT_WORK, "work_id": ask["work_id"], "data": results, "order": ask["order"]}), _master.address)
-            else:
-                raise AskTypeNotFoundError(f"No such type {ask['type']}")
-        except AskTypeNotFoundError as e:
-            logging.info(f"No such ask exists: {e}")
-            return
-        except WorkError as e:
-            logging.info(f"Encountered work error: {e}")
-            err = {"type": _WORK_ERROR, "work_id": ask["work_id"], "error": e}
-            _send_message(_encode_dict(err), _master.address)
-        except Exception as e:
-            logging.info(f"Could not handle request: {e}")
-            logging.info(traceback.format_exc())
-            return
-
-    def do_work(self, ask: Dict) -> List:
-        """After recieving work from the master server, execute the function on the data
-
-        :param ask: dict of type, function, array
-        :type ask: Dict
-        :return: list of computed data
-        :rtype: List
-        """
-        func = ask["function"]
-        data = ask["array"]
-
-        try:
-            results = list(map(func, data))
-        except Exception as e:
-            raise WorkError(f"Could not compute: {e} \n {traceback.format_exc()}")
-
-        return results
-
-
-class _ThreadedWorkerServer(socketserver.ThreadingMixIn, socketserver.TCPServer):
-    pass
+from ._worker import (ThreadedWorkerServer, WorkerServer,
+                      close_connection_with_master, reset_globals)
 
 
 class Worker:
@@ -102,10 +27,7 @@ class Worker:
                             level=logging.INFO)
         self.name = name
         self._server = None
-
-        global _master, _id
-        _master = None
-        _id = None
+        reset_globals()
 
     def start(self, ip: str = "localhost", port: int = 0) -> None:
         """Start the server on the given ip and port
@@ -117,8 +39,11 @@ class Worker:
         """
         self.__init__(self.name)
         address = (ip, port)
-        self._server = _ThreadedWorkerServer(address, _WorkerServer)
 
+        if self._server:
+            raise ServerAlreadyStartedError()
+
+        self._server = ThreadedWorkerServer(address, WorkerServer)
         t = threading.Thread(target=self._server.serve_forever, daemon=True)
         t.start()
         logging.info(f"Worker server running on {self.get_address()}")
@@ -128,9 +53,7 @@ class Worker:
         if self._server is None:
             logging.warning("No server has been started")
         else:
-            msg = _encode_dict({"type": _CLOSE_CONNECTION, "id": _id})
-            if _master:
-                _send_message(msg, _master.address)
+            close_connection_with_master()
             self._server.socket.close()
             self._server.shutdown()
             logging.info("Worker shutdown")
