+++ tests/utils.py
@@ -1,7 +1,7 @@
 import socket
 from typing import Tuple
-from overkill.utils.server_messaging_standards import _ACCEPT_WORK, _DELEGATE_WORK, _FINISHED_TASK, _NEW_CONNECTION
-from overkill.utils.utils import _decode_message, _encode_dict, _recv_msg, _send_message, _socket_send_message
+from overkill.servers.utils.server_messaging_standards import _ACCEPT_WORK, _DELEGATE_WORK, _FINISHED_TASK, _NEW_CONNECTION
+from overkill.servers.utils.utils import _decode_message, _encode_dict, _recv_msg, _send_message, _socket_send_message
 
 
 class MockWorker:
